backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
openai==0.27.8
python-multipart==0.0.6
websockets==11.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
